# backend/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Fight Club API",
    "description": "Complete API documentation for the AI Fight Club application - a platform for AI coding challenges and evaluation",
    "version": "1.0.0",
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    },
    "contact": {
      "name": "Support",
      "url": "https://aifightclub.example.com",
      "email": "support@aifightclub.example.com"
    },
    "termsOfService": "https://aifightclub.example.com/terms",
    "x-api-versioning": {
      "strategy": "uri-path",
      "current": "v1",
      "deprecated": [],
      "sunset": []
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current server (relative path)"
    }
  ],
...
  "paths": {
...
    "/users/{id}": {
      "get": {
        "summary": "Get user by ID",
        "description": "Retrieves a user by their unique identifier",
        "operationId": "getUserById",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to retrieve",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
    "/user-journey/events": {
      "get": {
        "summary": "Get user journey events",
        "description": "Retrieves all events in the current user's journey",
        "operationId": "getUserJourneyEvents",
        "tags": [
          "UserJourney"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by event type",
            "schema": {
              "type": "string",
              "enum": [
                "login",
                "challenge_started",
                "challenge_completed",
                "evaluation_received",
                "badge_earned",
                "level_up"
              ]
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "description": "Filter events from this date (inclusive)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "description": "Filter events to this date (inclusive)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User journey events retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserJourneyEvent"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationParams"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/user-journey/timeline": {
      "get": {
        "summary": "Get user timeline",
        "description": "Retrieves a chronological timeline of the user's journey",
        "operationId": "getUserTimeline",
        "tags": [
          "UserJourney"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "name": "includeDetails",
            "in": "query",
            "description": "Include detailed information about each event",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User timeline retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": {
                            "type": "string",
                            "format": "date",
                            "description": "Date of the events"
                          },
                          "events": {
                            "type": "array",
                            "items": {
                              "$ref": "#/components/schemas/UserJourneyEvent"
                            }
                          }
                        }
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationParams"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication. Use the token from the login response. Tokens expire after 24 hours."
      }
    },
    "parameters": {
      "limitParam": {
        "name": "limit",
        "in": "query",
        "description": "Maximum number of items to return per page (default: 20, max: 100)",
        "schema": {
          "type": "integer",
          "default": 20,
          "minimum": 1,
          "maximum": 100
        }
      },
      "offsetParam": {
        "name": "offset",
        "in": "query",
        "description": "Number of items to skip (for pagination)",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication information is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized - Invalid or missing token"
                }
              }
            }
          }
        }
      },
      "NotFoundError": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "Resource not found"
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "PaginationParams": {
        "type": "object",
        "description": "Parameters for pagination",
        "properties": {
          "limit": {
            "type": "integer",
            "default": 20,
            "minimum": 1,
            "maximum": 100,
            "description": "Maximum number of items to return"
          },
          "offset": {
            "type": "integer",
            "default": 0,
            "minimum": 0,
            "description": "Number of items to skip"
          },
          "sort": {
            "type": "string",
            "description": "Field to sort by (prefix with - for descending order)"
          }
        }
      },
...
      "User": {
        "type": "object",
        "description": "User information returned by the API",
        "required": [
          "id",
          "email"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the user"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user"
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the user"
          },
          "professionalTitle": {
            "type": "string",
            "description": "Professional title or position"
          },
          "location": {
            "type": "string",
            "description": "User's location"
          },
          "country": {
            "type": "string",
            "description": "User's country"
          },
          "focusArea": {
            "type": "string",
            "description": "Primary focus area or specialty"
          },
          "lastActive": {
            "type": "string",
            "format": "date-time",
            "description": "When the user was last active"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the user was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the user was last updated"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "suspended",
              "pending"
            ],
            "description": "Current user status",
            "default": "active"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "user",
                "admin"
              ]
            },
            "description": "Roles assigned to the user",
            "default": [
              "user"
            ]
          },
          "onboardingCompleted": {
            "type": "boolean",
            "description": "Whether the user has completed onboarding",
            "default": false
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the user is active",
            "readOnly": true
          },
          "hasCompletedProfile": {
            "type": "boolean",
            "description": "Whether the user has completed their profile",
            "readOnly": true
          },
          "preferences": {
            "type": "object",
            "description": "User preferences",
            "additionalProperties": true
          }
        }
      },
...
      "UserJourneyEvent": {
        "type": "object",
        "description": "Event in a user's journey",
        "required": [
          "id",
          "userId",
          "type",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the event"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the user the event belongs to"
          },
          "type": {
            "type": "string",
            "enum": [
              "login",
              "challenge_started",
              "challenge_completed",
              "evaluation_received",
              "badge_earned",
              "level_up"
            ],
            "description": "Type of event"
          },
          "data": {
            "type": "object",
            "description": "Additional data for the event",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the event occurred"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}